refactor(taste): extract shuffle weighting into a service module
ADR-0005 decision point 3 says to reuse listing.py's play-count/recency/favourite
weighting for the radio's taste term rather than reimplementing it. But that code
lives in a route module, so a service importing it would invert the dependency.

Moves ShufflePresetWeights, SHUFFLE_PRESETS, compute_track_weight and
apply_artist_variety to app/services/taste_weighting.py verbatim; listing.py
imports them back. The two helpers lose their leading underscore since they now
have callers outside their original module.

Pure movement, no logic changes — kept as its own commit so the behaviour-
preserving part is reviewable separately from the scoring work that follows.

The module docstring records the thing that matters for the next step:
compute_track_weight returns an unbounded multiplier (1.0 baseline, up to 4x per
axis across three axes plus a 3x favourites boost), not a 0-1 score, so anything
combining it with a normalised score must scale it first.

1175 tests pass; the 3 failures are the pre-existing chat contract tests that
assume no Anthropic key is configured.
